Discard previous renaming
